+++ src/app/App.tsx
@@ -1,7 +1,7 @@
 import { Component } from 'react';
 import { SearchPage } from '../pages/SearchPage/SearchPage.tsx';
 import './styles/variables/global.scss';
-import './app.css';
+import './app.scss';
 
 class App extends Component {
   render() {
