+++ src/components/List/List.tsx
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import style from './List.module.css';
+import style from './List.module.scss';
 import type { HeroResponse } from '../../models';
 
 interface ListProps {
