+++ src/components/Loader/Loader.tsx
@@ -1,6 +1,6 @@
 import { Component } from 'react';
 import loader from '../../assets/loader.png';
-import style from './Loader.module.css';
+import style from './Loader.module.scss';
 
 class Loader extends Component {
   render() {
