+++ src/pages/SearchPage/SearchPage.module.scss
